+++ fuzz/fuzzers/roundtrip.rs
@@ -2,10 +2,10 @@
 #[macro_use] extern crate libfuzzer_sys;
 extern crate base64;
 
-use base64::engine::DEFAULT_ENGINE;
+use base64::{Engine as _, engine::DEFAULT_ENGINE};
 
 fuzz_target!(|data: &[u8]| {
-    let encoded = base64::encode_engine(data, &DEFAULT_ENGINE);
+    let encoded = DEFAULT_ENGINE.encode(data);
     let decoded = base64::decode_engine(&encoded, &DEFAULT_ENGINE).unwrap();
     assert_eq!(data, decoded.as_slice());
 });