+++ tests/encode.rs
@@ -1,9 +1,7 @@
-use base64::alphabet::URL_SAFE;
-use base64::engine::general_purpose::PAD;
-use base64::*;
+use base64::{alphabet::URL_SAFE, engine::general_purpose::PAD, engine::DEFAULT_ENGINE, *};
 
 fn compare_encode(expected: &str, target: &[u8]) {
-    assert_eq!(expected, encode(target));
+    assert_eq!(expected, DEFAULT_ENGINE.encode(target));
 }
 
 #[test]
@@ -55,6 +53,6 @@ fn encode_all_bytes_url() {
          -AgYKDhIWGh4iJiouMjY6PkJGSk5SVlpeYmZqbnJ2en6ChoqOkpaanqKmqq6ytrq\
          -wsbKztLW2t7i5uru8vb6_wMHCw8TFxsfIycrLzM3Oz9DR0tPU1dbX2Nna29zd3t_g4eLj5OXm5-jp6uvs7e7v8PHy\
          8_T19vf4-fr7_P3-_w==",
-        encode_engine(&bytes, &engine::GeneralPurpose::new(&URL_SAFE, PAD),)
+        &engine::GeneralPurpose::new(&URL_SAFE, PAD).encode(&bytes)
     );
 }